Added Totals for Deductions, Wages Earned, and Hours.
Changed all occurences of String.Convert to String.Format respectively.

+++ PaycheckSim/WorkDay.cs
@@ -7,8 +7,8 @@ namespace PaycheckSim
     public class WorkDay
     {
         public string Day { get; set; }
-        public int totalHours { get; set; }
+        public double totalHours { get; set; }
         public int regHours { get; set; }
-        public int otHours { get; set; }
+        public double otHours { get; set; }
     }
 }
